Api Security Starter - 统一 ApiSecurity 注解属性，使一个注解支持多种接口安全功能

+++ robin-saas-dc/robin-api-security-spring-boot-starter/src/main/java/cn/com/mockingbird/robin/api/autoconfigure/ApiSecurityProperties.java
@@ -17,9 +17,9 @@ public class ApiSecurityProperties {
      */
     private Boolean enable = true;
     /**
-     * 签名过期时间
+     * 签名过期时间，单位是秒，默认 30s
      */
-    private Integer signatureExpiredTime = 30;
+    private Long signatureExpiredTime = 30L;
     /**
      * RSA 私钥
      */
